exercise-3/requirements.txt
Flask==3.0.3
Flask-Testing==0.8.1
psycopg2-binary==2.9.6